fix #1
`npm install` fails in linux (debian 10 buster, node v12.13), as reported in issue anvaka#1. 

Setting the flags as in this change fixes this problem by negating the default flag - which disable exception handling - , as per the solution described in [this issue](nodejs/node-gyp#17). 

It is good to have [this recommended solution](https://github.com/nodejs/node-gyp/issues/17#issuecomment-3917578) in mind:

> The real solution however would be to remove all instances of C++-style throw statements. That's what I did for node-ffi, but I'm sure that won't be possible for every module.